Fix /questions page refreshing on add/edit
